Remove Projectile kill method
