+++ Assets/Code/Unity/Tests/Actor/CardEntities/ActorTests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6599dc3048bf4024e89bebe829ecb5f8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
